+++ include/beast/websocket/detail/mask.hpp
@@ -19,6 +19,99 @@ namespace beast {
 namespace websocket {
 namespace detail {
 
+/** XOR-shift Generator.
+
+    Meets the requirements of UniformRandomNumberGenerator.
+    Simple and fast RNG based on:
+    http://xorshift.di.unimi.it/xorshift128plus.c
+    does not accept seed==0
+*/
+class xor_shift_engine
+{
+public:
+    using result_type = std::uint64_t;
+
+    xor_shift_engine(xor_shift_engine const&) = default;
+    xor_shift_engine& operator=(xor_shift_engine const&) = default;
+
+    explicit
+    xor_shift_engine(result_type val = 1977u)
+    {
+        seed(val);
+    }
+
+    void
+    seed(result_type seed);
+
+    template<class Sseq>
+    typename std::enable_if<
+        ! std::is_integral<Sseq>::value>::type
+    seed(Sseq& ss)
+    {
+        std::array<result_type, 2> v;
+        ss.generate(v.begin(), v.end());
+        seed((v[0] << 32) + v[1]);
+    }
+
+    result_type
+    operator()();
+
+    static
+    result_type constexpr
+    min()
+    {
+        return std::numeric_limits<result_type>::min();
+    }
+
+    static
+    result_type constexpr
+    max()
+    {
+        return std::numeric_limits<result_type>::max();
+    }
+
+private:
+    result_type s_[2];
+
+    static
+    result_type
+    murmurhash3(result_type x);
+};
+
+inline
+void
+xor_shift_engine::seed(result_type seed)
+{
+    if(seed == 0)
+        throw std::domain_error("invalid seed");
+    s_[0] = murmurhash3(seed);
+    s_[1] = murmurhash3(s_[0]);
+}
+
+inline
+auto
+xor_shift_engine::operator()() ->
+    result_type
+{
+    result_type s1 = s_[0];
+    result_type const s0 = s_[1];
+    s_[0] = s0;
+    s1 ^= s1<< 23;
+    return(s_[1] =(s1 ^ s0 ^(s1 >> 17) ^(s0 >> 26))) + s0;
+}
+
+inline
+auto
+xor_shift_engine::murmurhash3(result_type x)
+    -> result_type
+{
+    x ^= x >> 33;
+    x *= 0xff51afd7ed558ccdULL;
+    x ^= x >> 33;
+    x *= 0xc4ceb9fe1a85ec53ULL;
+    return x ^= x >> 33;
+}
+
 // Pseudo-random source of mask keys
 //
 template<class Generator>
@@ -27,8 +120,7 @@ class maskgen_t
     Generator g_;
 
 public:
-    using result_type =
-        typename Generator::result_type;
+    using result_type = std::uint32_t;
 
     maskgen_t();
 
@@ -51,7 +143,7 @@ maskgen_t<Generator>::operator()() noexcept ->
     result_type
 {
     for(;;)
-        if(auto key = g_())
+        if(auto key = static_cast<result_type>(g_()))
             return key;
 }
 
@@ -59,15 +151,20 @@ template<class _>
 void
 maskgen_t<_>::rekey()
 {
-    std::random_device rng;
-    std::array<std::uint32_t, 32> e;
-    for(auto& i : e)
-        i = rng();
-    std::seed_seq ss(e.begin(), e.end());
-    g_.seed(ss);
+    for(;;)
+    {
+        std::random_device rng;
+        std::array<std::uint32_t, 32> e;
+        for(auto& i : e)
+            i = rng();
+        std::seed_seq ss(e.begin(), e.end());
+        g_.seed(ss);
+        if(g_())
+            break;
+    }
 }
 
-using maskgen = maskgen_t<std::mt19937>;
+using maskgen = maskgen_t<xor_shift_engine>;
 
 //------------------------------------------------------------------------------